bugfix: avatar_updated from int -> double + set mode to light (in info.plist)

+++ wftiles/Player.swift
@@ -13,7 +13,7 @@ class Player: NSObject, Codable, NSCoding {
     let score: Int
     let id: UInt64
     var rack: [String]?
-    let avatar_updated: UInt64?
+    let avatar_updated: Double?
     let fb_first_name: String?
     let fb_middle_name: String?
     let fb_last_name: String?
@@ -29,7 +29,7 @@ class Player: NSObject, Codable, NSCoding {
         static let fb_last_name = "fb_last_name"
     }
     
-    init(username: String, score: Int, id: UInt64, rack: [String]?, avatar_updated: UInt64?, fb_first_name: String?, fb_middle_name: String?, fb_last_name: String?) {
+    init(username: String, score: Int, id: UInt64, rack: [String]?, avatar_updated: Double?, fb_first_name: String?, fb_middle_name: String?, fb_last_name: String?) {
         self.username = username
         self.score = score
         self.id = id
@@ -57,7 +57,7 @@ class Player: NSObject, Codable, NSCoding {
         let score = aDecoder.decodeInteger(forKey: PropertyKey.score)
         let id = aDecoder.decodeObject(forKey: PropertyKey.id) as! UInt64
         let rack = aDecoder.decodeObject(forKey: PropertyKey.rack) as? [String]
-        let avatar_updated = aDecoder.decodeObject(forKey: PropertyKey.avatar_updated) as? UInt64
+        let avatar_updated = aDecoder.decodeObject(forKey: PropertyKey.avatar_updated) as? Double
         let fb_first_name = aDecoder.decodeObject(forKey: PropertyKey.fb_first_name) as? String
         let fb_middle_name = aDecoder.decodeObject(forKey: PropertyKey.fb_middle_name) as? String
         let fb_last_name = aDecoder.decodeObject(forKey: PropertyKey.fb_last_name) as? String
